# openapi_v2.yaml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /api/flashcards/{id}/:
    get:
      operationId: retrieveFlashcard
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - api
    put:
      operationId: updateFlashcard
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flashcard'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Flashcard'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Flashcard'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - api
    patch:
      operationId: partialUpdateFlashcard
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flashcard'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Flashcard'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Flashcard'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - api
    delete:
      operationId: destroyFlashcard
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      responses:
        '204':
          description: ''
      tags:
      - api
  /api/flashcards/{id}/retention/:
    get:
      operationId: getRetentionFlashcard
      description: Get current retention probability for a card.
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - api
...
  /generation/flashcards/{flashcard_set_id}/export-anki/:
    get:
      operationId: listexport_flashcard_set_to_ankis
      description: Export a specific flashcard set to Anki format (.apkg file).
      parameters:
      - name: flashcard_set_id
        in: path
        required: true
        description: ''
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: {}
          description: ''
      tags:
      - generation
...
  /generation/api/flashcards/{id}/:
    get:
      operationId: retrieveFlashcard
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - generation
    put:
      operationId: updateFlashcard
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flashcard'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Flashcard'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Flashcard'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - generation
    patch:
      operationId: partialUpdateFlashcard
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flashcard'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Flashcard'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Flashcard'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - generation
    delete:
      operationId: destroyFlashcard
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      responses:
        '204':
          description: ''
      tags:
      - generation
  /generation/api/flashcards/{id}/retention/:
    get:
      operationId: getRetentionFlashcard
      description: Get current retention probability for a card.
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - generation
...
  /api/flashcards/{id}/reset/:
    post:
      operationId: resetCardFlashcard
      description: Reset a card to new state (useful for relearning).
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flashcard'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Flashcard'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Flashcard'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - api
  /api/flashcards/{id}/review/:
    post:
      operationId: reviewCardFlashcard
      description: "Review a single flashcard and update its schedule.\n\nRequest\
        \ body:\n{\n    \"quality\": 3,  // 0-5 quality rating\n    \"response_time_seconds\"\
        : 15.5  // optional\n}"
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this flashcard.
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flashcard'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Flashcard'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Flashcard'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flashcard'
          description: ''
      tags:
      - api
...
  /generation/flashcards/generate/:
    post:
      operationId: creategenerate_flashcards_view
      description: ''
      parameters: []
      requestBody:
        content:
          application/json:
            schema: {}
          application/x-www-form-urlencoded:
            schema: {}
          multipart/form-data:
            schema: {}
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: ''
      tags:
      - generation
...
  /generation/flashcards/export-multiple-anki/:
    post:
      operationId: createexport_multiple_flashcard_sets_to_anki
      description: Export multiple flashcard sets to a single Anki deck.
      parameters: []
      requestBody:
        content:
          application/json:
            schema: {}
          application/x-www-form-urlencoded:
            schema: {}
          multipart/form-data:
            schema: {}
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: ''
      tags:
      - generation
...
components:
  schemas:
...
    Flashcard:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        question:
          type: string
        answer:
          type: string
        updated_at:
          type: string
          format: date-time
          readOnly: true
        flashcard_set:
          type: integer
      required:
      - question
      - answer
      - flashcard_set
...